Dismiss Card
Dismissing the card with an external event (e.g. button click) is
available again. Removed unused code.
